third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/focus-navigation-negative-tabindex-expected.txt

CONSOLE ERROR: line 87: Uncaught TypeError: element.createShadowRoot is not a function
This tests navigation from non keyboard-focusable shadow host to its shadow. crbug.com/446584

FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
